+++ controllers/comment.controller.js
@@ -20,7 +20,7 @@ class CommentController {
       const { userKey } = res.locals.user;
 
       if (userKey == 0) {
-        res.status(400).send({ message: "로그인이 필요합니다." });
+        return res.status(400).send({ message: "로그인이 필요합니다." });
       }
 
       if (!comment) {
@@ -44,6 +44,10 @@ class CommentController {
       const { comment } = req.body;
       const { userKey } = res.locals.user;
 
+      if (userKey == 0) {
+        return res.status(400).send({ message: "수정권한이 없습니다." });
+      }
+
       const updateComment = await this.commentService.updateComment(
         userKey,
         commentId,
@@ -64,6 +68,10 @@ class CommentController {
     try {
       const { commentId } = req.params;
       const { userKey } = res.locals.user;
+
+      if (userKey == 0) {
+        return res.status(400).send({ message: "삭제권한이 없습니다." });
+      }
       const deleteComment = await this.commentService.deleteComment(
         commentId,
         userKey
@@ -85,7 +93,7 @@ class CommentController {
       const { userKey } = res.locals.user;
 
       if (userKey == 0) {
-        res.status(400).send({ message: "로그인이 필요합니다." });
+        return res.status(400).send({ message: "로그인이 필요합니다." });
       }
 
       const Likes = await this.commentService.updateCommentLike(
